LDBR-1.11: Переместить интерфейсы в отдельный модуль, переместить модуль main в папку cmd/api/
