+++ repositories/stores/session_test.go
@@ -1,4 +1,4 @@
-package repositories
+package stores
 
 import (
 	"backendServer/models"
